Namespace custom simple_form input fields
