+++ app/inputs/archangel/wysiwyg_input.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Archangel
+  # WYSIWYG custom simple_form input field
+  #
+  # @author dfreerksen
+  # @since 0.0.1
+  #
+  class WysiwygInput < SimpleForm::Inputs::TextInput
+  end
+end
