--- app/inputs/wysiwyg_input.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-# WYSIWYG custom simple_form input field
-#
-# @author dfreerksen
-# @since 0.0.1
-#
-class WysiwygInput < SimpleForm::Inputs::TextInput
-end
