+++ config/initializers/simple_form.rb
@@ -135,7 +135,7 @@ SimpleForm.setup do |config|
 
   # Namespaces where SimpleForm should look for custom input classes that
   # override default inputs.
-  # config.custom_inputs_namespaces << "CustomInputs"
+  config.custom_inputs_namespaces << "Archangel"
 
   # Default priority for time_zone inputs.
   # config.time_zone_priority = nil
